+++ crates/apollo-compiler/src/diagnostics.rs
@@ -1,5 +1,5 @@
 use crate::ast::DirectiveLocation;
-use crate::validation::FileId;
+use crate::diagnostic::CliReport;
 use crate::validation::NodeLocation;
 use std::fmt;
 use thiserror::Error;
@@ -245,31 +245,13 @@ pub(crate) enum DiagnosticData {
 }
 
 impl ApolloDiagnostic {
-    pub(crate) fn to_report(
-        &self,
-        config: ariadne::Config,
-    ) -> ariadne::Report<'static, NodeLocation> {
-        use ariadne::{ColorGenerator, Report, ReportKind};
-
-        let mut colors = ColorGenerator::new();
-        let (id, offset) = if let Some(location) = self.location {
-            (location.file_id(), location.offset())
-        } else {
-            (FileId::NONE, 0)
-        };
-        let mut builder = Report::build(ReportKind::Error, id, offset)
-            .with_config(config)
-            .with_message(&self.data);
-        builder.add_labels(self.labels.iter().filter_map(|label| {
-            label.location.map(|loc| {
-                ariadne::Label::new(loc)
-                    .with_message(&label.text)
-                    .with_color(colors.next())
-            })
-        }));
-        if let Some(help) = &self.help {
-            builder = builder.with_help(help);
+    pub(crate) fn report(&self, report: &mut CliReport) {
+        report.with_message(&self.data);
+        for label in &self.labels {
+            report.with_label_opt(label.location, &label.text);
         }
-        builder.finish()
+        if let Some(help) = &self.help {
+            report.with_help(help);
+        }
     }
 }